requirements.txt
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.28.0
fastapi>=0.110.0
uvicorn[standard]>=0.23.2
sqlmodel>=0.0.9
jsonschema>=4.20.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
SQLAlchemy==2.0.23
google-api-python-client>=2.108.0
google-auth>=2.23.0
google-auth-oauthlib>=1.2.0
python-socketio[asyncio]>=5.11.0
psycopg[binary]>=3.1
psycopg2-binary>=2.9
